Implement generation of shared secret for metadata
The ``nova-compute`` charm gates enablement of the local Nova
metadata service on presence of the ``metadata-shared-secret``
key in the relation data in the subordinate relation with
``neutron-plugin`` interface.  This is implemented in the classic
``neutron-openvswitch`` charm today.

Add the bits necessary for a reactive charm to participate in this
part of the relation.

Also adds units tests to the interface.

Change-Id: I358e60c40a7f7c0cd33b31f2b7a8c58d6cb83326